doc: handle elixir type info in from_map/1 & to_map/1
If a top-level key named "type": "Foo" is present, check to see if a
corresponding atom named :"Elixir.Foo" is present in the existing
atom table.

This permits user functions to dispatch based on atom type in the
returned doc struct, to an Ecto schema function directly, or similar
custom user function as required.

- upgrade from_map/1 and to_map/1
- add coerce_to_elixir_type/1
- add coerce_to_json_string/1
- remove struct related names from type fields
- sprinkle a few tests